.claude/skills/research-by-reddit/scripts/requirements.txt
# Reddit API
praw>=7.7.0

# Anthropic Claude API
anthropic>=0.40.0

# HTTP 请求
requests>=2.31.0

# 环境变量
python-dotenv>=1.0.0

# JSON 处理
orjson>=3.9.0

# 日期时间
python-dateutil>=2.8.2

# 进度条（可选）
tqdm>=4.66.0
